Add files via upload
Installation
To run this project locally, follow these steps:

Clone this repository using git clone https://github.com/vishesh-pandey/v-music.git
Install dependencies using npm install
Start the development server using npm startUsage
Once the development server is running, you can access the app by visiting http://localhost:3000 in your web browser. Here's how to use the app:

Enter a search term in the search bar and click the "Search" button
The app will fetch a list of tracks from the Spotify API and display them in a grid
Click the "Play" button on a track to play a 30-second preview of the song
Use the audio player controls to pause, resume, and adjust the volume of the song
Contributing
If you would like to contribute to this project, feel free to submit a pull request or open an issue.